+++ main/src/main/java/net/oneandone/stool/util/Processes.java
@@ -19,60 +19,102 @@ import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.file.FileNode;
 import net.oneandone.sushi.launcher.Failure;
 import net.oneandone.sushi.launcher.Launcher;
+import net.oneandone.sushi.util.Separator;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Processes {
-    public static Processes create(World world) throws Failure {
-        return new Processes(new Launcher(world.getWorking(), "ps", "aux").exec());
+    // group 1: pid
+    // group 2: pcpu
+    // group 3: pmem
+    // group 4: command
+    private static final Pattern PS_LINE = Pattern.compile("^(\\d+)\\s+(\\S+)\\s+(\\S+)\\s+(.*)$");
+
+    public static Processes load(World world) throws Failure {
+        return create(new Launcher(world.getWorking(), "ps", "ax", "-o", "pid=,pcpu=,pmem=,command=").exec());
     }
 
-    // group 1: user
-    // group 2: process id
-    // group 3: rest of line
-    private static final Pattern PS_AUX_LINE = Pattern.compile("^(\\S+)\\s+(\\d+)\\s+(.+)$", Pattern.MULTILINE);
+    public static Processes create(String str) throws Failure {
+        Processes result;
+        Matcher matcher;
+        int pid;
+        double cpu;
+        double mem;
+        String command;
 
-    private final String ps;
+        result = new Processes();
+        for (String line : Separator.RAW_LINE.split(str)) {
+            line = line.trim();
+            matcher = PS_LINE.matcher(line);
+            if (!matcher.matches()) {
+                throw new IllegalArgumentException(line);
+            }
+            pid = Integer.parseInt(matcher.group(1));
+            cpu = Double.parseDouble(matcher.group(2));
+            mem = Double.parseDouble(matcher.group(3));
+            command = matcher.group(4);
+            result.add(new Data(pid, cpu, mem, command));
+        }
+        return result;
+    }
 
-    public Processes(String ps) {
-        this.ps = ps;
+    public static class Data {
+        public final int pid;
+        public final double cpu;
+        public final double mem;
+        public final String command;
+
+        public Data(int pid, double cpu, double mem, String command) {
+            this.pid = pid;
+            this.cpu = cpu;
+            this.mem = mem;
+            this.command = command;
+        }
+
+        public String toString() {
+            return pid + " " + cpu + " " + mem + " " + command;
+        }
+    }
+
+    //--
+
+    private final Map<String, Data> all;
+
+    public Processes() {
+        all = new HashMap<>();
+    }
+
+    public void add(Data data) {
+        if (all.put(Integer.toString(data.pid), data) != null) {
+            throw new IllegalArgumentException(data.toString());
+        }
+    }
+
+    public Data lookup(String pid) {
+        return all.get(pid);
     }
 
     public boolean hasPid(String pid) {
-        Matcher matcher;
-
-
-        matcher = PS_AUX_LINE.matcher(ps);
-        while (matcher.find()) {
-            if (pid.equals(matcher.group(2))) {
-                return true;
-            }
-        }
-        return false;
+        return all.containsKey(pid);
     }
 
-    public String tomcatPid(FileNode stageWrapper) {
-        Matcher matcher;
-        String result;
-        String tmp;
+    public String servicePid(FileNode backstage) {
         String key;
-        String remaining;
+        int result;
 
-        matcher = PS_AUX_LINE.matcher(ps);
-        result = null;
-        key = "wrapper.statusfile=" + stageWrapper.join("shared/run/tomcat.status").getAbsolute();
-        while (matcher.find()) {
-            remaining = matcher.group(3);
-            if (remaining.contains(key)) {
-                tmp = matcher.group(2);
-                if (result == null) {
-                    result = tmp;
-                } else {
-                    throw new IllegalStateException("ambiguous: " + result + " vs " + matcher.group(2));
+        result = 0;
+        key = "wrapper.statusfile=" + backstage.join("shared/run/tomcat.status").getAbsolute();
+        for (Data data : all.values()) {
+            if (data.command.contains(key)) {
+                if (result != 0) {
+                    throw new IllegalStateException("ambiguous: " + result + " vs " + data.pid);
                 }
+                result = data.pid;
             }
         }
-        return result;
+        return result == 0 ? null : Integer.toString(result);
     }
 }